pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tripclaim-auto-approval"
version = "0.1.0"
description = "Trip claim folder auto-compliance approval"
readme = ""
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "pypdf>=4.2.0",
  "python-docx>=1.1.2",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "streamlit>=1.36.0",
  "requests>=2.32.0",
]

[project.scripts]
tripclaim = "tripclaim.main:app"
tripclaim-api = "tripclaim.api:run"
tripclaim-ui = "tripclaim.ui:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
